Stats: Add pagination to tribe conquests.

+++ sql/stats/players/conquests-loss-count.sql
@@ -1,3 +1,3 @@
-SELECT COUNT(*)
+SELECT COUNT(*)::int
 FROM ${worldId:name}.conquests
 WHERE old_owner = ${playerId}
